% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_expression}
\alias{assert_expression}
\title{Vérifie la conformité d'une expression}
\usage{
assert_expression(expr, .var.name)
}
\arguments{
\item{expr}{une expression à évaluer}

\item{.var.name}{Nom de l'objet à vérifier pour afficher dans les messages}
}
\value{
En sortie la fonction retourne l'objet \code{x} (le résultat de l'évaluation de l'expression \code{expr}) de manière invisible ou une erreur.
}
\description{
Vérifie la conformité d'une expression
}
\details{
La fonction évalue l'expression \code{expr}. Si elle génère une erreur ou un warning, alors on retourne une erreur. Si elle ne génère aucun message particulier, on retourne alors l'objet \code{x} (le résultat de l'évaluation de l'expression \code{expr}), sans erreur.
}
\examples{

assert_expression(expr = {2 + 2})
assert_expression(expr = {is.integer(1L)})
try(assert_expression(expr = {log("a")}), silent = TRUE)

}
