% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_frequency}
\alias{assert_frequency}
\title{Vérifie la conformité d'une fréquence}
\usage{
assert_frequency(x, add = NULL, .var.name = checkmate::vname(x))
}
\arguments{
\item{x}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}

\item{add}{Collection pour stocker les messages d'erreurs (Default is NULL)}

\item{.var.name}{Nom de l'objet à vérifier pour afficher dans les messages}
}
\value{
En sortie la fonction retourne l'objet \code{x} de manière invisible ou une erreur.
}
\description{
Vérifie la conformité d'une fréquence
}
\details{
La fréquence d'une série temporelle est soit mensuelle (\code{12L} ou \code{12.}) soit trimestrielle (\code{4L} ou \code{4.}). Les autres fréquences ne sont pas acceptées.
Cette fonction s'appuie essentiellement sur les fonctions \code{checkmate::assert_int} et \code{checkmate::assert_choice}.
Il y a néanmoins une petite subtilité : on vérifie si l'objet \code{x} est de type double ou integer.
Dans le premier cas, on affichera un warning et on corrigera l'objet au format integer pour les traitements ultérieurs. En sortie, \code{x} est retourné de manière invisible.
}
\examples{

assert_frequency(4L)
assert_frequency(12L)

}
