% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_scalar_integer}
\alias{assert_scalar_integer}
\title{Vérifie la conformité d'un entier scalaire}
\usage{
assert_scalar_integer(
  x,
  add = NULL,
  .var.name = checkmate::vname(x),
  warn = TRUE
)
}
\arguments{
\item{x}{un entier}

\item{add}{Collection pour stocker les messages d'erreurs (Default is NULL)}

\item{.var.name}{Nom de l'objet à vérifier pour afficher dans les messages}

\item{warn}{un booleen}
}
\value{
En sortie la fonction retourne l'objet \code{x} de manière invisible ou une erreur.
}
\description{
Vérifie la conformité d'un entier scalaire
}
\details{
On vérifie que l'objet \code{x} en entrée est bien un entier.
Cette fonction s'appuie essentiellement sur la fonction \code{checkmate::assert_int}.
Il y a néanmoins une petite subtilité : on vérifie si l'objet \code{x} est de type double ou integer.
Dans le premier cas, on affichera un warning et on corrigera l'objet au format integer pour les traitements ultérieurs. En sortie, \code{x} est retourné de manière invisible.
Si l'argument \code{warn} est \code{FALSE}, alors la fonction ne retournera pas de warning lors de l'évaluation.
}
\examples{

assert_scalar_integer(1L)
assert_scalar_integer(100L)
assert_scalar_integer(-4L)
assert_scalar_integer(0L)

}
\seealso{
\code{\link[=assert_scalar_natural]{assert_scalar_natural()}}
}
