% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_scalar_natural}
\alias{assert_scalar_natural}
\title{Vérifie la conformité d'un entier naturel}
\usage{
assert_scalar_natural(x, add = NULL, .var.name = checkmate::vname(x))
}
\arguments{
\item{x}{un entier naturel (strictement positif)}

\item{add}{Collection pour stocker les messages d'erreurs (Default is NULL)}

\item{.var.name}{Nom de l'objet à vérifier pour afficher dans les messages}
}
\value{
En sortie la fonction retourne l'objet \code{x} de manière invisible ou une erreur.
}
\description{
Vérifie la conformité d'un entier naturel
}
\details{
Cette fonction s'appuie essentiellement sur la fonction \code{checkmate::assert_count}. Il y a néanmoins une petite subtilité : on vérifie si l'objet est de type double ou integer. Dans le premier cas, on affichera un warning et on corrigera l'objet pour les traitements ultérieurs puis l'objet au format integer sera retourné de manière invisible.
}
\examples{

# Avec des entier integer
assert_scalar_natural(1L)
assert_scalar_natural(100L)

# Avec des entiers double
assert_scalar_natural(2.)
assert_scalar_natural(457)

}
