% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_TimeUnits}
\alias{assert_TimeUnits}
\title{Vérifie la conformité d'un objet TimeUnits}
\usage{
assert_TimeUnits(x, frequency_ts, add = NULL, .var.name = checkmate::vname(x))
}
\arguments{
\item{x}{un numérique qui représente le time units de}

\item{frequency_ts}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}

\item{add}{Collection pour stocker les messages d'erreurs (Default is NULL)}

\item{.var.name}{Nom de l'objet à vérifier pour afficher dans les messages}
}
\value{
En sortie la fonction retourne l'objet \code{x} de manière invisible ou une erreur.
}
\description{
Vérifie la conformité d'un objet TimeUnits
}
\details{
Un objet de type TimeUnits est un numérique qui désigne l'année et la période en cours avec ses décimales.
Ainsi pour une série temporelle mensuelle, \code{2020.5} représente la moitié de l'année donc juillet 2020 et s'écrit \code{c(2020L, 7L)} au format date_ts.
}
\examples{

assert_TimeUnits(2020.5, frequency_ts = 12L)
assert_TimeUnits(2020.5, frequency_ts = 4L)
assert_TimeUnits(2023, frequency_ts = 12L)

assert_TimeUnits(2000 + 5/12, frequency_ts = 12L)
assert_TimeUnits(2015 + 3/4, frequency_ts = 4L)

}
