% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{combine2ts}
\alias{combine2ts}
\title{Combiner 2 ts}
\usage{
combine2ts(a, b)
}
\arguments{
\item{a}{un objet ts}

\item{b}{un objet ts}
}
\value{
En sortie, la fonction retourne la combinaison des 2 objets.
}
\description{
La fonction \code{combine2ts} permet de combiner (comme c()) 2 time series.
}
\examples{
#La temporalité du second ts n'est pas prise en compte
x1 <- ts(rep(1, 4), start = 2000, frequency = 12)
x2 <- ts(rep(2, 4), start = 2020, frequency = 12)
combine2ts(x1, x2)

#Attention aux NA présent en début ou fin de ts !
ev_pib |> combine2ts(x1)
}
