% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{date_ts2TimeUnits}
\alias{date_ts2TimeUnits}
\title{Conversion d'une date du format date_ts au format TimeUnits}
\usage{
date_ts2TimeUnits(date_ts, frequency_ts)
}
\arguments{
\item{date_ts}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{frequency_ts}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}
}
\value{
En sortie, la fonction retourne la date au format AAAA + TT/4 ou AAAA + MM/12.
}
\description{
Conversion d'une date du format date_ts au format TimeUnits
}
\details{
AAAA signifie que l'année est au format numérique avec 4 chiffres (Exemple : l'année deux mille vingt-deux s'écrit 2022 et non 22)
MM signifie que le mois est au format numérique (Exemple : le mois de mai s'écrit 5, le moi de décembre s'écrit 12)
TT signifie que le trimestre est au format numérique (Exemple : le troisième trimestre s'écrit 3)
}
\examples{

date_ts2TimeUnits(date_ts = c(2020L, 4L), frequency_ts = 12L) # Avril 2020
date_ts2TimeUnits(date_ts = c(2022L, 11L), frequency_ts = 12L) # Novembre 2020

date_ts2TimeUnits(date_ts = c(2022, 4L), frequency_ts = 4L) # 4ème trimestre de 2022
date_ts2TimeUnits(date_ts = c(1995L, 2L), frequency_ts = 4L) # 2ème trimestre de 1995

}
