% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{getTimeUnits}
\alias{getTimeUnits}
\title{Conversion d'une date TS au format TimeUnits}
\usage{
getTimeUnits(date, frequency)
}
\arguments{
\item{date}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{frequency}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.)}
}
\value{
En sortie, la fonction retourne la date au format AAAA + TT/4 ou AAAA + MM/12.
}
\description{
Conversion d'une date TS au format TimeUnits
}
