% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_vector}
\alias{is_vector}
\title{Vérifie qu'on objet est bien un vecteur non vide}
\usage{
is_vector(x)
}
\arguments{
\item{x}{un vecteur}
}
\value{
En sortie la fonction retourne un booleen qui précise si l'argument \code{x} est bien un vecteur (unidimentionnel) non vide.
}
\description{
Vérifie qu'on objet est bien un vecteur non vide
}
\examples{

# Réponse positive
is_vector(1:3)
is_vector(letters)
is_vector(seq(
    from = as.Date("2000-01-01"),
    to = as.Date("2023-05-28"),
    by = "month"))

# Réponse négative
is_vector(NULL)
is_vector(list(1, 2, 3))
is_vector()
}
