% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{isGoodTS}
\alias{isGoodTS}
\title{Vérifie la conformité d'un objet ts dans le cadre des enquêtes de conjoncture}
\usage{
isGoodTS(dataTS, withWarning = TRUE)
}
\arguments{
\item{dataTS}{un objet ts unidimensionnel}

\item{withWarning}{un booléen}
}
\value{
En sortie la fonction retourne un booleen qui précise si le TS est conforme ou non.
Dans le cas où withWarning vaut TRUE et que le TS n'est pas conforme, un warning qui précise la raison sera déclenché.
}
\description{
Vérifie la conformité d'un objet ts dans le cadre des enquêtes de conjoncture
}
\details{
Les fonctions du package ts4conj sont faites pour fonctionner avec des times-series de fréquence mensuelle ou trimestrielle et basés sur le système des mois, trimestres et années classiques. On travaille avec des données numériques (integer, double ou logical) mais les autres types atomic sont acceptés également.
}
\examples{
ts1 <- ts(1:100, start = 2010L, frequency = 12L)
ts2 <- ts(1:100, start = 2010 + 1/7, frequency = 12L)
ts3 <- ts(1:100, start = 2010L, frequency = 1L)

isGoodTS(ts1)
isGoodTS(ts2)
isGoodTS(ts3)

isGoodTS(ts2, withWarning = FALSE)
isGoodTS(ts3, withWarning = FALSE)
}
