% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{isTSdate}
\alias{isTSdate}
\title{Vérifie le format de date}
\usage{
isTSdate(date)
}
\arguments{
\item{date}{un vecteur numérique, de préférence integer}
}
\value{
En sortie la fonction retourne un booleen et un warning additionnel si besoin.
}
\description{
La fonction \code{isTSdate} vérifie qu'un objet est de type AAAA, c(AAAA, MM) ou c(AAAA, TT)
}
\details{
Les fonctions du package ts4conj sont faites pour fonctionner avec des times-series de fréquence mensuelle ou trimestrielle et basé sur le système des mois, trimestres et années classiques.
On cherche donc à favoriser l'utilisation de vecteur c(AAAA, MM) pour désigner la date choisie.
}
\examples{
#De bons formats de date
isTSdate(c(2020L, 4L))
isTSdate(2022L)

#Formats avec un warning
isTSdate(c(2020, 4))
isTSdate(2022)

#Format non accepté --> erreur
isTSdate(2022.5)
isTSdate(2022 + 1/12)
}
