% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{lastDate}
\alias{lastDate}
\title{Dernière date non NA}
\usage{
lastDate(dataTS)
}
\arguments{
\item{dataTS}{un objet ts unidimensionnel conforme aux règles de assert_ts}
}
\value{
En sortie, la fonction retourne un objet au format \code{date_ts} (\code{AAAA}, \code{c(AAAA, MM)} ou \code{c(AAAA, TT)})
}
\description{
Cette fonction calcule la dernière date pour laquelle l'objet \code{dataTS} ne vaut pas NA.
}
\details{
La date retournée en output est au format \code{date_ts}. Si l'objet \code{dataTS} ne contient que des NAs, la fonction retourne une erreur.
}
\examples{

ts1 <- ts(c(NA, NA, NA, 1:10, NA), start = 2000, frequency = 12)
ts2 <- ts(c(1:10), start = 2020, frequency = 4)

stats::end(ts1)
lastDate(ts1)

stats::end(ts1)
lastDate(ts2)

}
\seealso{
\code{firstDate}
}
