% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{mens2trim}
\alias{mens2trim}
\title{Correspondance d'une date mensuelle au trimestre courant}
\usage{
mens2trim(date_ts)
}
\arguments{
\item{date_ts}{un vecteur numérique, de préférence \code{integer}, au format c(AAAA, MM)}
}
\value{
En sortie, la fonction retourne la date au format c(AAAA, TT) correspondant au trimestre courant du mois \code{date_ts}.
}
\description{
La fonction \code{mens2trim} prend en argument une date au format c(AAAA, MM) et la convertit au format c(AAAA, TT) avec le trimestre en cours.
}
\examples{
mens2trim(c(2019L, 4L)) #Avril 2019 --> 2ème trimestre 2019
mens2trim(c(2020L, 11L)) #Novembre 2020 --> 4ème trimestre 2020

}
