% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{nextDate}
\alias{nextDate}
\title{Obtenir la date suivante}
\usage{
nextDate(date, frequency, lag = 1L)
}
\arguments{
\item{date}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{frequency}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}

\item{lag}{un entier}
}
\value{
En sortie, la fonction retourne un vecteur d'entier qui représente la date à la période future.
}
\description{
Obtenir la date suivante
}
\details{
Lorsqu'on parle de date suivante, on parle de date future.
L'argument \code{lag} est entier et désigne le nombre de décalage que l'on affecte à notre date.
Par exemple pour des lag positif (1L, 2L, 10L) on désigne le décalage de la période suivante, celle d'après et celle dans 10 périodes.
Cependant, lorsque l'argument \code{lag} vaut zéro, la fonction retourne la \code{date} inchangée. Aussi lorsque l'argument \code{lag} est négatif, la fonction se comporte comme la fonction \code{previousDate} et retourne les périodes passées et non futures.
}
\examples{

nextDate(c(2020L, 4L), frequency = 4L, lag = 2L)
nextDate(c(2021L, 1L), frequency = 4L, lag = -2L)

nextDate(c(2020L, 4L), frequency = 12L, lag = 2L)
nextDate(c(2022L, 6L), frequency = 12L, lag = 12L)
}
\seealso{
\code{previousDate}
}
