% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{previousDate}
\alias{previousDate}
\title{Obtenir la date précédente}
\usage{
previousDate(date_ts, frequency, lag = 1L)
}
\arguments{
\item{frequency}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}

\item{lag}{un entier}

\item{date}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}
}
\value{
En sortie, la fonction retourne un vecteur d'entier qui représente la date à la période passée.
}
\description{
Obtenir la date précédente
}
\details{
Lorsqu'on parle de date précédente, on parle de date passée.
L'argument \code{lag} est entier et désigne le nombre de décalage que l'on affecte à notre date.
Par exemple pour des lag positif (1L, 2L, 10L) on désigne le décalage de la période précédente, celle d'avant et celle d'il y a 10 périodes.
Cependant, lorsque l'argument \code{lag} vaut zéro, la fonction retourne la \code{date} inchangée. Aussi lorsque l'argument \code{lag} est négatif, la fonction se comporte comme la fonction \code{nextDate} et retourne les périodes futures et non passées.
}
\examples{

previousDate(c(2020L, 4L), frequency = 4L, lag = 2L)
previousDate(c(2021L, 1L), frequency = 4L, lag = -2L)

previousDate(c(2020L, 4L), frequency = 12L, lag = 2L)
previousDate(c(2022L, 6L), frequency = 12L, lag = 12L)
}
\seealso{
\code{nextDate}
}
