% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{setValue_ts}
\alias{setValue_ts}
\title{Change certaines valeurs d'un ts}
\usage{
setValue_ts(dataTS, date_ts, value)
}
\arguments{
\item{dataTS}{un objet ts unidimensionnel conforme aux règles de isGoodTS}

\item{date_ts}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{value}{un vecteur de même type que le ts \code{dataTS}}
}
\value{
En sortie, la fonction retourne l'objet \code{dataTS} n objet ts modifié avec les valeurs de \code{value} imputés à partir de la date \code{date}.
}
\description{
La fonction \code{setValue_ts} modifie la ou les valeurs d'un objet ts à une date donnée.
}
\examples{
ev_pib |> setValue_ts(date = c(2021L, 2L), value = c(1, 2, 3))
}
