% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TractorTsbox-package.R
\docType{package}
\name{TractorTsbox-package}
\alias{TractorTsbox}
\alias{TractorTsbox-package}
\title{TractorTsbox: Outils de gestion des ts et des dates des ts}
\description{
Ce package propose des outils pour gérer les dates des objets ts. On utilise que des dates mensuelles ou trimestrielles sous la forme c(2022, 9) pour septembre 2022 ou c(2021, 2) pour le deuxième trimestre 2021 selon la fréquence. L'idée générale est de se passer des dates 2021-07-01 ou 2021.5833333 qui sont soit trop contraignantes, soit trop imprécises selon les arrondies. On cherche à trouver un compromis entre simplicité, précision et utilisation des fonctions stats de base de création et de gestion des ts (ts, window).
}
\author{
\strong{Maintainer}: Tanguy Barthelemy \email{tangbarth@hotmail.fr}

}
\keyword{internal}
