% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{trim2mens}
\alias{trim2mens}
\title{Correspondance d'une date trimestrielle au premier mois du trimestre}
\usage{
trim2mens(date_ts)
}
\arguments{
\item{date_ts}{un vecteur numérique, de préférence \code{integer}, au format \code{c(AAAA, TT)}}
}
\value{
En sortie, la fonction retourne la date au format \code{date_ts} (c'est-à-dire un vecteur d'entiers de la forme \code{c(AAAA, MM)})
}
\description{
La fonction \code{trim2mens} prend en argument une date au format \code{c(AAAA, TT)} et la convertit au format \code{c(AAAA, MM)} en choisissant le premier mois du trimestre.
}
\examples{
trim2mens(c(2019L, 4L)) #4ème trimestre 2019 --> Octobre 2019
trim2mens(c(2020L, 1L)) #1er trimestre 2020 --> Janvier 2020

}
