% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ts2df}
\alias{ts2df}
\title{Convertit un objet \code{ts} en \code{data.frame}}
\usage{
ts2df(x)
}
\arguments{
\item{x}{un objet de type \code{ts}.}
}
\value{
En sortie la fonction retourne un \code{data.frame} avec autant de
colonnes que \code{x} et une de plus pour la date.
}
\description{
Convertit un objet \code{ts} en \code{data.frame}
}
\examples{

ts2df(AirPassengers)

}
