% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.YYYYMM}
\alias{as.YYYYMM}
\title{Conversion au format c(AAAA, MM)}
\usage{
as.YYYYMM(TimeUnits)
}
\arguments{
\item{TimeUnits}{une date en année (Par exemple 2021.83333333333 pour Novembre 2021)}
}
\value{
En sortie, la fonction retourne la date au format c(AAAA, MM)
}
\description{
La fonction \code{as.YYYYMM} convertit une date en année en date au format c(AAAA, MM).
}
\examples{
as.YYYYMM(2019.75) #Octobre 2019
as.YYYYMM(2020) #Janvier 2020
as.YYYYMM(2020 + 1/12) #Février 2020
as.YYYYMM(2020 + 12/12) #Janvier 2021
}
