% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.YYYYTT}
\alias{as.YYYYTT}
\title{Conversion au format c(AAAA, TT)}
\usage{
as.YYYYTT(TimeUnits)
}
\arguments{
\item{TimeUnits}{une date en année (Par exemple 2015.25 pour le 2ème trimestre 2015)}
}
\value{
En sortie, la fonction retourne la date au format (AAAA, TT)
}
\description{
La fonction \code{as.YYYYTT} convertit une date en année en date au format c(AAAA, TT).
}
\examples{
as.YYYYTT(2019.75) #4ème trimestre 2019
as.YYYYTT(2020) #1er trimestre 2020
as.YYYYTT(2022 + 1/4) #2ème trimestre 2022
}
