% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{combine2ts}
\alias{combine2ts}
\title{Combiner 2 ts}
\usage{
combine2ts(a, b)
}
\arguments{
\item{a}{un objet ts}

\item{b}{un objet ts}
}
\value{
En sortie, la fonction retourne un ts qui contient les valeurs de \code{a} aux temps de \code{a} et les valeurs de \code{b} aux temps de \code{b}.
}
\description{
La fonction \code{combine2ts} combine (comme c()) 2 time series de même fréquence (mensuelle ou trimestrielle).
}
\details{
Si \code{a} et \code{b} ont une période en commun, les valeurs de \code{b} écrasent celles de a sur la période concernée.
Si il existe une période sur laquelle ni \code{a} ni \code{b} ne prennent de valeur (mais qu'il existe des valeurs à des dates ultérieures et antérieures) alors le ts en sortie prendra NA sur cette période.
}
\examples{
#La temporalité du second ts n'est pas prise en compte
x1 <- stats::ts(rep(1, 4), start = 2000, frequency = 12)
x2 <- stats::ts(rep(2, 4), start = 2020, frequency = 12)
combine2ts(x1, x2)

#Attention aux NA présent en début ou fin de ts !
ev_pib |> combine2ts(x1)
}
