% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getValue_ts}
\alias{getValue_ts}
\title{Récupère des valeurs d'un ts}
\usage{
getValue_ts(dataTS, date_ts, n = 1L)
}
\arguments{
\item{dataTS}{un objet ts unidimensionnel conforme aux règles de assert_ts}

\item{date_ts}{un vecteur numérique, de préférence \code{integer} au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{n}{un entier}
}
\value{
En sortie, la fonction retourne un vecteur avec les valeurs extraites.
}
\description{
La fonction \code{getValue_ts} permet de récupérer des valeurs.
}
\examples{

ts1 <- ts(1:10, start = 2023, frequency = 12)
ts2 <- ts(letters, start = 2020, frequency = 4)

getValue_ts(ts1, date_ts = c(2023L, 7L))
getValue_ts(ts2, date_ts = c(2021L, 4L), n = 4L)

}
