% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{isGoodDate}
\alias{isGoodDate}
\title{Vérifie le format de date}
\usage{
isGoodDate(date)
}
\arguments{
\item{date}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}
}
\value{
En sortie la fonction retourne un booleen et un warning additionnel si besoin.
}
\description{
La fonction \code{isGoodDate} vérifie qu'un objet est de type AAAA, c(AAAA, MM) ou c(AAAA, TT)
}
\details{
Les fonctions du package ts4conj sont faites pour fonctionner avec des times-series de fréquence mensuelle ou trimestrielle et basés sur le système des mois, trimestres et années classiques.
On cherche donc à favoriser l'utilisation de vecteur c(AAAA, MM) pour désigner la date choisie.
}
\examples{
#De bons formats de date
isGoodDate(c(2020L, 4L))
isGoodDate(2022L)

#Formats avec un warning
isGoodDate(c(2020, 4))
isGoodDate(2022)

#Format non accepté --> erreur
isGoodDate(2022.5)
isGoodDate(2022 + 1/12)
}
