% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelles.R
\name{libelles}
\alias{libelles}
\title{Libelés pour une période}
\usage{
libelles(date_ts, frequency, nb = 1)
}
\arguments{
\item{date_ts}{un vecteur numérique, de préférence integer au format AAAA, c(AAAA, MM) ou c(AAAA, TT)}

\item{frequency}{un entier qui vaut 4L (ou 4.) pour les séries trimestrielles et 12L (ou 12.) pour les séries mensuelles.}

\item{nb}{un entier}
}
\value{
En sortie, la fonction retourne un vecteur de char avec les libelés de la période.
}
\description{
La fonction \code{libelles} créé un vecteur de chaines de caractère contenant les libelés de toutes les dates sur une période
}
\details{
Pour choisir la période, il faut spécifier une date de début de période et un nombre de valeur.
}
\examples{
libelles(date_ts = c(2019L, 10L), frequency = 12L, nb = 9L)
libelles(date_ts = c(2019L, 4L), frequency = 4L, nb = 3L)
}
